Added a couple of comments to the script
This is where an extended description would appear! I created a new branch and am now committing directly to that branch. I see now that I could have directly edited the master and at this point created a new branch for the commit and started a pull request. After I commit these changes, I will start a pull request (which will start the process of us going back and forth prior to accepting the changes in the master - I think!)